+++ panda_pytest_assertions/assert_context/__init__.py
@@ -6,10 +6,11 @@ pytest.register_assert_rewrite(__name__ + '._assert_context')
 # isort: on
 
 
-from ._assert_context import assert_context, AssertContext
+from ._assert_context import assert_context, AssertContext, ExceptionMatch
 
 
 __all__ = [
     'assert_context',
     'AssertContext',
+    'ExceptionMatch',
 ]