+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * `assert_context` module
     * `exception` parameter accepts `None` value which behaves exactly the same as if the value was not set at all
+    * `exception` accepts `ExceptionMatch` instances which allow specifying not only exception type, but also the value for match parameter for `pytest.raises`
 
 ## [0.1.1] - 2024-11-13